refactor: structuring user-preferences the same as in mP codebase
