+++ src/services/user-preferences/composite-user-preferences-service/index.ts
@@ -1,0 +1,1 @@
+export * from './composite-user-preferences-service'
