+++ src/services/user-preferences/models/definitions/user-preference-definition.ts
@@ -1,0 +1,6 @@
+import { type UserPreferenceScopeType } from './user-preference-scope-type'
+
+export interface UserPreferenceDefinition {
+  isOptedInByDefault: boolean
+  allowedScope: UserPreferenceScopeType
+}
