+++ src/services/user-preferences/models/definitions/user-preference-definitions.ts
@@ -1,0 +1,5 @@
+import { type UserPreferenceDefinition } from './user-preference-definition'
+
+export type UserPreferenceDefinitions<UserPreferenceId extends PropertyKey> = {
+  [Id in UserPreferenceId]?: UserPreferenceDefinition
+}
