+++ src/services/user-preferences/models/definitions/user-preference-scope-type.ts
@@ -1,0 +1,6 @@
+export enum UserPreferenceScopeType {
+  Global,
+  LevelOneScope,
+  LevelTwoScope,
+  LevelThreeScope,
+}
