+++ src/services/user-preferences/models/storage-models/user-preference.ts
@@ -1,0 +1,3 @@
+export interface UserPreference {
+  optedIn: boolean
+}
