+++ src/services/user-preferences/models/storage-models/user-preferences.ts
@@ -1,0 +1,6 @@
+import { type UserPreferenceScope } from './user-preference-scope'
+import { type UserPreference } from './user-preference'
+
+export type UserPreferences<UserPreferenceId extends PropertyKey> = {
+  [K in UserPreferenceScope]?: { [Id in UserPreferenceId]: UserPreference }
+}
