+++ src/services/user-preferences/models/user-preferences/composite-user-preferences.ts
@@ -1,0 +1,5 @@
+import { type UserPreference } from '../storage-models/user-preference'
+
+export type CompositeUserPreferences<UserPreferenceId extends PropertyKey> = {
+  [Id in UserPreferenceId]: UserPreference
+}
